Migrate web DOM actions to Svelte attachments
