+++ apps/web/src/lib/attachments/row-activate.ts
@@ -1,0 +1,37 @@
+import type { Attachment } from 'svelte/attachments';
+
+const MOVE_THRESHOLD = 6;
+
+export function rowActivate(getOnActivate: () => () => void): Attachment<HTMLElement> {
+	return (node) => {
+		let downX = 0;
+		let downY = 0;
+
+		function handlePointerDown(event: PointerEvent) {
+			downX = event.clientX;
+			downY = event.clientY;
+		}
+
+		function handleClick(event: MouseEvent) {
+			if (Math.hypot(event.clientX - downX, event.clientY - downY) >= MOVE_THRESHOLD) return;
+			getOnActivate()();
+		}
+
+		function handleKeyDown(event: KeyboardEvent) {
+			if (event.key === 'Enter' || event.key === ' ') {
+				event.preventDefault();
+				getOnActivate()();
+			}
+		}
+
+		node.addEventListener('pointerdown', handlePointerDown);
+		node.addEventListener('click', handleClick);
+		node.addEventListener('keydown', handleKeyDown);
+
+		return () => {
+			node.removeEventListener('pointerdown', handlePointerDown);
+			node.removeEventListener('click', handleClick);
+			node.removeEventListener('keydown', handleKeyDown);
+		};
+	};
+}
