+++ apps/web/src/lib/components/log/InlineLogRow.svelte
@@ -4,7 +4,7 @@
 	import { formatLogRowTimestamp } from '$lib/utils/time';
 	import { getByPath } from '$lib/utils/get-by-path';
 	import { formatCell } from '$lib/utils/column-width';
-	import { rowActivate } from '$lib/actions/row-activate';
+	import { rowActivate } from '$lib/attachments/row-activate';
 
 	let {
 		hit,
@@ -35,7 +35,7 @@
 	tabindex="0"
 	class="border-line relative block min-h-[25px] py-1 pr-2 pl-3 text-left font-mono text-xs before:absolute before:top-[1px] before:bottom-[1px] before:left-0 before:w-[3px] before:bg-[var(--level-color)] before:content-[''] hover:bg-[color-mix(in_oklab,var(--level-color)_14%,transparent)] {layout}"
 	style="--level-color: {levelColor(hit.level)};"
-	use:rowActivate={onActivate}
+	{@attach rowActivate(() => onActivate)}
 >
 	{#each parts as part, i (i)}{#if i > 0}<span class="px-2" aria-hidden="true">|</span
 			>{/if}{part}{/each}
